requirements.txt
absl-py==1.1.0
astunparse==1.6.3
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chex==0.1.3
contextlib2==21.6.0
cryptography @ file:///tmp/build/80754af9/cryptography_1652101588893/work
cycler==0.11.0
dm-tree==0.1.7
etils==0.6.0
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.2
grpcio==1.46.3
h5py==3.7.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.11.4
importlib-resources==5.7.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
lmdb==1.3.0
Markdown==3.3.7
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
msgpack==1.0.3
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
oauthlib==3.2.0
opt-einsum==3.3.0
optax==0.1.2
packaging==21.3
Pillow==9.0.1
promise==2.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pygame==2.1.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
pytorch-fid==0.2.1
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1656438147783/work
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.8.0
termcolor==1.1.0
toolz==0.11.2
tqdm==4.64.0
typing_extensions==4.2.0
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650639997961/work
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
git+https://github.com/jk4011/torchquad.git@multiple_integrands
Cython